requirements.txt
scikit_image==0.16.2
matplotlib==3.1.3
tensorflow==2.1.0
numpy==1.18.1
Keras==2.3.1
Pillow==7.1.1
